fix: GBIF integration issues — search, images, taxonomy, distributions
- Fix case-sensitive species search by normalizing queries to Title Case
- Fix Wikipedia binomial regex matching common names before scientific names
- Fix GBIF images not used: occurrence key overflow (int→long), vision-less
  screening model now defaults to SAFE for curated GBIF images
- Fix "Global" appearing in native region by filtering broad locality values
- Fix CoL API response parsing (data nested under usage.name, not top-level)
- Fix taxonomy class/order duplication when GBIF has no order level
- Store CoL taxonId, authorship, and synonyms in taxonomy entity

+++ src/Creaturedex.AI/Models/GbifAnimalData.cs
@@ -53,7 +53,7 @@ public record GbifImageResult
     public string? RightsHolder { get; init; }
     public string? Publisher { get; init; }
     public string? Country { get; init; }
-    public int? GbifOccurrenceId { get; init; }
+    public long? GbifOccurrenceId { get; init; }
     public string? MediaIdentifierMd5 { get; init; }
     public string CachedUrl => MediaIdentifierMd5 != null && GbifOccurrenceId != null
         ? $"https://api.gbif.org/v1/image/cache/occurrence/{GbifOccurrenceId}/media/{MediaIdentifierMd5}"
